Create BookListSection component and connect it with main logic
