+++ script/index.js
@@ -1,36 +1,44 @@
-import { setupAddBookForm, bookList } from '../components/AddNewBookSection.js';
+// This script is the entry point for the book management application.
+import { Book, BookList } from '../modules/Book.js';
 
-// Function to render the list of books
+// Import the renderBooks function from BookListSection
+import { renderBooks } from '../components/BookListSection.js';
 
-const renderBooks = () => {
-  const listContainer = document.getElementById('book-list');
-  listContainer.innerHTML = ''; // Clear existing content
+// Creating an instance of BookList
+const bookList = new BookList();
 
-  bookList.getBooks().forEach((book) => {
-    const bookElement = document.createElement('div');
-    bookElement.className = 'book-entry';
+// Add a sample book to the list
 
-    const text = document.createElement('p');
-    text.textContent = `"${book.title}" by ${book.author}`;
-
-    const removeBtn = document.createElement('button');
-    removeBtn.textContent = 'Remove';
-    removeBtn.className = 'remove-btn';
-
-    removeBtn.addEventListener('click', () => {
-      bookList.removeBook(book.id);
-      renderBooks();
-    });
-
-    bookElement.appendChild(text);
-    bookElement.appendChild(removeBtn);
-    listContainer.appendChild(bookElement);
-  });
+const addNewBook = (title, author) => {
+  const newBook = new Book(title, author);
+  bookList.addBook(newBook);
+  renderBooks(bookList); // Re-render the book list after adding a new book
 };
 
-// Initial rendering of books
-renderBooks();
-setupAddBookForm(renderBooks);
+// Rendering the books when the page loads
+
+renderBooks(bookList);
+
+// Function to handle adding a new book from a form
+const form = document.getElementById('book-form');
+
+form.addEventListener('submit', (e) => {
+  e.preventDefault(); // Prevent the default form submission
+
+  const titleInput = document.getElementById('title');
+  const authorInput = document.getElementById('author');
+
+  const title = titleInput.value.trim();
+  const author = authorInput.value.trim();
+
+  if (title && author) {
+    addNewBook(title, author); // Add the new book
+    form.reset(); // Reset the form fields
+  } else {
+    alert('Please fill in both fields.'); // Alert if fields are empty
+  }
+});
+
